refactor(domain.Payment) : amount 필드의 타입을 long 으로 변경

+++ src/main/java/lotto/domain/Payment.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
 public class Payment {
-    private final int amount;
+    private final long amount;
 
     public Payment(int amount) {
         this.amount = amount;
